+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Developing_Code/Old_Arm_Claw/_2023_10_27_05_Arm_Claw_V5.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Developing_Code;
+package org.firstinspires.ftc.teamcode.Developing_Code.Old_Arm_Claw;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -10,9 +10,9 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
-@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "V1 Arm Claw")
+@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "V5 Arm Claw")
 @Disabled
-public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
+public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
         DcMotor frontLeftMotor = hardwareMap.dcMotor.get("motorFL");
@@ -24,6 +24,7 @@ public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
         Servo droneServo = hardwareMap.servo.get("drone");
         Servo clawLeft = hardwareMap.servo.get("clawLeft");
         Servo clawRight = hardwareMap.servo.get("clawRight");
+        Servo axle = hardwareMap.servo.get("axle");
 
         frontRightMotor.setDirection(DcMotorSimple.Direction.FORWARD);
         backRightMotor.setDirection(DcMotorSimple.Direction.FORWARD);
@@ -43,15 +44,16 @@ public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
 
         // Adjust the orientation parameters to match your robot
         IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.LEFT,
+                RevHubOrientationOnRobot.LogoFacingDirection.RIGHT,
                 RevHubOrientationOnRobot.UsbFacingDirection.UP));
         // Without this, the REV Hub's orientation is assumed to be logo up / USB backward
         imu.initialize(parameters);
         droneServo.scaleRange(0, 1);
-        droneServo.setPosition(0.85);
 
         clawLeft.scaleRange(0, 1);
         clawRight.scaleRange(0, 1);
+        
+        axle.scaleRange(0, 1);
 
         //clawRight.setPosition(0);
         //clawLeft.setPosition(0);
@@ -61,6 +63,11 @@ public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
         double closeClaw = 0.2; //decrease to close more
 
         waitForStart();
+
+        axle.setPosition(0.8);
+        droneServo.setPosition(0.7);
+        clawLeft.setPosition(openClaw);
+        clawRight.setPosition(openClaw);
 
         while (opModeIsActive()) {
             //gamepad 1 - drive base
@@ -97,39 +104,64 @@ public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
             double liftPower = arm.getPower();
 
             if (gamepad1.right_trigger > 0) {
-                liftTargetPosition += 5;
+                liftTargetPosition += 2.5;
             } else if (gamepad1.left_trigger > 0) {
-                liftTargetPosition -= 5;
+                liftTargetPosition -= 2.5;
             }
-            if (liftTargetPosition > 967) {
-                liftTargetPosition = 962;
+            if (liftTargetPosition > 700) {
+                liftTargetPosition = 695;
             }
-            if (liftTargetPosition < 0) {
-                liftTargetPosition = 0;
+            if (liftTargetPosition < 5) {
+                liftTargetPosition = 5;
             }
 
             if (gamepad1.right_bumper) {
-                liftTargetPosition = 962;
+                for (int liftPos = 0; liftPos <= 500; liftPos = liftPos + 5) {
+                    liftTargetPosition = liftPos;
+                    arm.setTargetPosition(liftTargetPosition);
+                    arm.setPower(0.7);
+                    arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+                    sleep(15);
+                }
+
+                clawLeft.setPosition(openClaw);
+                clawRight.setPosition(openClaw);
+                
+                sleep(300);
+                axle.setPosition(0);
             }
             if (gamepad1.left_bumper) {
-                liftTargetPosition = 0;
+                liftTargetPosition = 10;
+                arm.setTargetPosition(liftTargetPosition);
+                arm.setPower(0.1);
+                arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+                
+                axle.setPosition(0.8);
             }
 
-            arm.setTargetPosition(liftTargetPosition);
-            arm.setPower(1);
-            arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            if (arm.isBusy() == false) {
+                arm.setTargetPosition(liftTargetPosition);
+                arm.setPower(0.7);
+                arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+            }
 
             double clawLeftTarget = 0;
-
-            //wrist code
 
             if (gamepad1.b) {
                 clawLeft.setPosition(closeClaw);
                 clawRight.setPosition(closeClaw);
+                axle.setPosition(0.8);
             }
-            if (gamepad1.y) {
+            if (gamepad1.a) {
                 clawLeft.setPosition(openClaw);
                 clawRight.setPosition(openClaw);
+            }
+
+            if (gamepad1.dpad_up) {
+                axle.setPosition(0);
+            }
+            if (gamepad1.dpad_down) {
+                axle.setPosition(0.8);
             }
 
             telemetry.addData("Claw Left Target: ", clawLeftTarget);
@@ -137,10 +169,10 @@ public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
 
             double droneServoPosition = droneServo.getPosition();
 
-            if (gamepad1.a) {
+            if (gamepad1.y) {
                 droneServo.setPosition(1);
                 sleep(1500);
-                droneServo.setPosition(0.85);
+                droneServo.setPosition(0.7);
             }
 
             // ADDED CODE - sends info about current servo position to driver station
@@ -148,7 +180,8 @@ public class _2023_10_27_01_Arm_Claw_V1 extends LinearOpMode {
 
             telemetry.addData("Claw Left Position: ", clawLeft.getPosition());
             telemetry.addData("Claw Right Position: ", clawRight.getPosition());
-            //telemetry.addData("Claw Target: ", claw);555 iirj dc
+
+            telemetry.addData("Axle Position:", axle.getPosition());
 
             telemetry.addData("Arm position: ", arm.getCurrentPosition());
             telemetry.addData("Arm power: ", liftPower);
